% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pandas_statement.R
\name{create_pandas_statement}
\alias{create_pandas_statement}
\title{Create a chained pandas command string}
\usage{
create_pandas_statement(
  df_name,
  filter_str = NULL,
  select_str = NULL,
  sort_by_str = NULL,
  sort_asc_str = NULL,
  assing_str = NULL
)
}
\arguments{
\item{df_name}{A character string for the name of the pandas DataFrame.}

\item{filter_str}{A string for the \code{.query()} method.}

\item{select_str}{A string for column selection (e.g., "\link{'col1', 'col2'}").}

\item{sort_by_str}{A string for the \code{by} argument of \code{.sort_values()}.}

\item{sort_asc_str}{A string for the \code{ascending} argument of \code{.sort_values()}.}
}
\value{
A character string of the complete, chained pandas command.
}
\description{
This internal function assembles a Python command string for pandas by
chaining together different data manipulation methods. It serves as the
central command generator for the package.
}
\keyword{internal}
