% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_filter.R
\name{\%notin\%}
\alias{\%notin\%}
\title{"Not In" Operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Vector of values to be matched.}

\item{y}{Vector of values to be matched against.}
}
\value{
A logical vector.
}
\description{
Provides the opposite of the standard R \code{\%in\%} operator.
}
\examples{
"a" \%notin\% c("b", "c")
}
