% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_filter.R
\name{rp_mutate}
\alias{rp_mutate}
\title{Mutate (add/modify/remove) columns using pandas}
\usage{
rp_mutate(.data, to_remove = NULL, ..., return.as = "result")
}
\arguments{
\item{.data}{An R data frame.}

\item{to_remove}{A character vector of column names to remove.}

\item{...}{Named expressions for new/modified columns.}

\item{return.as}{Either "result", "code", or "all".}
}
\value{
A data frame or list depending on return.as.
}
\description{
Mutate (add/modify/remove) columns using pandas
}
