% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandas-package.R
\docType{package}
\name{rPandas-package}
\alias{rPandas-package}
\alias{rPandas}
\title{rPandas: A User-Friendly R Interface to Pandas}
\description{
This package provides a set of wrapper functions that allow R users to
interact with Python's pandas library using familiar R syntax.
}
\author{
\strong{Maintainer}: Akshat Maurya \email{codingmaster902@gmail.com}

Authors:
\itemize{
  \item Rihaan Satia \email{first.last@example.com}
  \item David Shilane \email{david.shilane@columbia.edu}
}

}
\keyword{package}
