% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_engine.R
\name{translate_assign_recursive}
\alias{translate_assign_recursive}
\title{Recursively translate an R expression for a pandas .assign() lambda}
\usage{
translate_assign_recursive(expr_body)
}
\arguments{
\item{expr_body}{A language object (call, symbol, or atomic).}
}
\value{
A character string of the translated Python expression.
}
\description{
Recursively translate an R expression for a pandas .assign() lambda
}
\keyword{internal}
