% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_filter.R
\name{translate_filter_recursive}
\alias{translate_filter_recursive}
\title{Recursive helper to translate R expressions}
\usage{
translate_filter_recursive(expr_body, env)
}
\arguments{
\item{expr_body}{A language object (call, symbol, or atomic).}
}
\value{
A character string.
}
\description{
Recursive helper to translate R expressions
}
\keyword{internal}
