% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_engine.R
\name{translate_groupby}
\alias{translate_groupby}
\title{Translate a .by argument into a pandas .groupby() string}
\usage{
translate_groupby(by_expr)
}
\arguments{
\item{by_expr}{An enquosured \code{.by} argument.}
}
\value{
A string for the \code{.groupby(..., as_index=False)} method,
or NULL if the argument is empty.
}
\description{
Translate a .by argument into a pandas .groupby() string
}
\keyword{internal}
