% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_engine.R
\name{translate_select}
\alias{translate_select}
\title{Translate captured column names into a Python list string}
\usage{
translate_select(...)
}
\arguments{
\item{...}{Bare column names (captured by \code{rlang::enquos}).}
}
\value{
A character string formatted as a Python list.
}
\description{
Translate captured column names into a Python list string
}
\keyword{internal}
