% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandas-package.R
\docType{_package}
\name{rPandas-package}
\alias{rPandas-package}
\title{rPandas: A User-Friendly R Interface to Pandas}
\description{
This package provides a set of wrapper functions that allow R users to
interact with Python's pandas library using familiar R syntax.
}
